Cleanup and simplify bots
Bots now answer to 'update'
They will be notified that the game is complete via the 'update' command as well.

Bet handling with the bots is handled by the game manager instead of directly
by the bet handler.

Players no longer need to know their position, it shouldn't matter to them
directly and can be assertained by the bet handler.

Players now have the 'name' set directly. This should be what allows bots
to keep track of each other, not a seat number.